Split main script into modules
